feat: export chalk and logger
The default logger instance is now exported for those whom wish to use it. It doesn't really affect Robo.js if it's used in weird ways anyway.

The `chalk` library has also been exported but not from the main package. This was mostly done for better compatibilty with default commands internally.