feat: automatic promise rejection handling
Robo.js will now automatically prevent unhandled promise rejections from crashing the process when:

- Running on production
- Running on dev with debug channel

You must set a debug channel for crashes to be prevented. I understand why Node behaves this way and I agree in that developers should not merely ignore such errors. That said, sometimes these errors are out of the developers' control and thus this is relevant.