feat: new automatic "dev" command when running in development mode
Development builds now have access to a new automatic command: "dev"

I'm trying to keep auto commands as minimal as possible to prevent bloat but this one has the potential to become extremely useful.

It's only available in development builds. Production builds exclude this completely.

We also require that a DISCORD_GUILD_ID environment variable is present to avoid accidental exposure to the public. This way, the command is only ever available to the developer and not actual users.

For now, it only shows logs the same way that error messages do. However, this can be used for many more useful things.