feat: manifest scopes/permissions & inherit from plugins
Permissions and scopes are now computed at the manifest level rather than when generating the invite. This allows us to inherit them from plugins and have them registered at build-time which can be handy for reviewing plugins.

I recognize that the current approach of computing permissions based on intents is not technically accurate, but it gets close enough for now. This will eventually be replaced with proper usage scanning.