refactor(cli): revamped internal manifest generation code
Not really much of a change for users to be aware of, more than the fact that it happened. This piece of code has been bothering me for a while and I finally decided to rewrite it.

generateObjectFromDirectory is now split into two functions: scanDir and generateEntries.

scanDir does what the name implies. It walks through a file tree and notifies a predicate function. All files run first before  nested directories (children) to prevent potential bugs.

The generateEntries now provides overloads for simpler usage (no more type casts) with only the entry logic. Conditionals were reduced.

We also now refer to CommandConfigs here as CommandEntries for consistency and flexibility.