feat: new "flashcore" api for easy persistent storage - compatible with keyv adapters
KV-based persistent storage compatible with `keyv` adapters!
https://github.com/jaredwray/keyv

The purpose of this is to have a common API available to the entire framework for persistent storage. It's still very much a work in progress and will improve. I'd love to be able to get this working with vector queries, for example.

The default implementation uses a primitive FS-based store. This should be good enough for small bots, but I highly recommend using a database adapter for higher volume.