refactor: avoid terminating workers unless necessary for graceful spirit shutdowns
Part of the reason why I added flushing support for the logger is that logs would sometimes go missing because the thread was terminated before logs were fully flushed.

By no longer terminating threads unless we need to, we prevent that + other similar issues. This also means that threads now correctly exit with code 0 instead of 1.