patch(spirits): exit worker threads with process.exit()
Now we not only finish communication and flush stuff ourselves, but also force an exit similar to how the non-spirits process works.

This is because, while it does work, the way Node's event loop works means that anything else added by developers that stays in the event loop indefinitely causes the threads to never die. The new method forces that to stop. Developers can still clean stuff up with `_stop` and `_restart` events.