feat: spirits system now enabled by default
This has been put to the test for a while now and has gone various improvements. So much that it might even be more stable than the older system.

The legacy process system is still available via a new `legacyProcess` experimental flag and will be retired in the next version after this one if all goes well.