fix(cli): only apply spirit system in buildasync function for non-plugin builds
The spirit system combined with the `build plugin` command previously caused issues because it would incorrectly try to execute the build via spirits rather than a forked bash process. This is more prominent now that it's the default.

We now check that a command was provided in order to apply it as a bash command.

It may make more sense to decoupled this into a separate function altogether in the future.