feat(plugin-ai): configurable commands and channel whitelists
To conserve token usage but still use command calls, it's now possible to pass an array of which commands to include. You can also pass `true` to include all commands.

You can now also pass an array of channels IDs to whitelist the bot in. That way, people won't need to mention it each time in such channels.

I've also cleaned up a bit of logging.