refactor(plugin-ai): new engine-based model usage
This still needs refinement... lots of it, but now it's possible to switch AI engines for this plugin! A lot of the Discord-specific logic has been abstracted away into the new `AI` object which replaces the old `AiEngine` object.

The actual options passed down to the underlying engine still need refinement, but at least it's possible now, and cleaner too!

Oh yeah, and we switched to path aliases because why not.