feat: support for embeds and files in natural language commands
Command results that contain files or embeds will now be sent correctly instead of ignored!

This is quite a big change, as we had to reword a lot of internals to support this. The `onReply` is also no longer part of engines. It is now part of the core `AI` module and the reply contains richer data.

We make a best effort to get GPT to correctly include files within the context in replies.