feat(plugin/ai): use assistant api when insight enabled (default: true)
We now use the wonderful new APIs made available to us by OpenAI to integrate long-term memory and data retrieval!

It's a somewhat complex process and currently does not work with function calls, but it is extremely helpful for data retrieval via the new `/documents` folder! I had to install a lightweight dependency for FormData as Node's built-in implementation had issues with OpenAI's API.