refactor(plugin/ai): clean up assistant files
We now auto delete files no longer in use by the assistant and keep a cache to track them to avoid duplicate uploads, thereby savings space by no longer unnecessarily uploading files that don't need to be!

Oh, and this also adds a validation check for directories. We won't be supporting those yet.