feat(plugin/ai): rewrite, custom models, assistant api, rich replies, end-user commands, bug fixes, smarter context (#160)
* refactor(plugin-ai): new engine-based model usage

This still needs refinement... lots of it, but now it's possible to switch AI engines for this plugin! A lot of the Discord-specific logic has been abstracted away into the new `AI` object which replaces the old `AiEngine` object.

The actual options passed down to the underlying engine still need refinement, but at least it's possible now, and cleaner too!

Oh yeah, and we switched to path aliases because why not.

* chore(plugin-ai): re-enabled /ai say changeset

* feat(plugin-ai): new `/chat` command

Direct replacement of the `/chat` command in the old GPT Plugin. This uses the core engine for responses, minus the fancy natural responses.

* feat(plugin-ai): image generation capabilities

* feat(plugin-ai): new `/imagine` command

* refactor(ai/engine): added function handlers, engine info, and init lifecycle functions

This will help us abstract away preparation steps, and soon functions as well!

The info will be used for debugging purposes later.

* refactor(ai): moved function loading logic to engine

This is very specific to OpenAI, so it makes sense to abstract unless we decide to abstract this very concept.

* refactor(ai): decoupled openai engine implementation from api calls

It helps to keep API logic separate, considering it is independent.

* refactor(ai): simplified api request logic

Retries and exponential backoff now built into a single function, which greatly simplifies other API calls!

I also fixed incorrect logger usage as well, added a smol utility.

* refactor(ai): re-imagined /imagine descriptions

Now more LLM-friendly and exciting!

* refactor!: removed `openaiKey` plugin option

Removed key input via plugin options as bad practice. The recommended way is now to have a specific environment variable, similar to how the framework treats Discord tokens.

* patch(ai): patch: stable command options with natural language

Commands no longer need to be optimized for this plugin! It will now make a much better effort towards getting the correct types and options now that we override more option functions.

* feat: support for embeds and files in natural language commands

Command results that contain files or embeds will now be sent correctly instead of ignored!

This is quite a big change, as we had to reword a lot of internals to support this. The `onReply` is also no longer part of engines. It is now part of the core `AI` module and the reply contains richer data.

We make a best effort to get GPT to correctly include files within the context in replies.

* feat: support for embeds and files in natural language commands

* refactor(ai): imagine command now accounts for multiple files

Not yet necessary, but useful in case multiple images were created.

* docs(ai): focus on env key over option, mention voice capabilities, and added hints

* feat(ai): support for components in natural language commands

Buttons with natural language? Yes please!

* feat(plugin/ai): use assistant api when insight enabled (default: true)

We now use the wonderful new APIs made available to us by OpenAI to integrate long-term memory and data retrieval!

It's a somewhat complex process and currently does not work with function calls, but it is extremely helpful for data retrieval via the new `/documents` folder! I had to install a lightweight dependency for FormData as Node's built-in implementation had issues with OpenAI's API.

* fix(plugin/ai): better array comparisons in `compare` helper function

By sorting arrays before comparing, we ensure we're only comparing its contents rather than the order of them as well!

* refactor(plugin/ai): clean up assistant files

We now auto delete files no longer in use by the assistant and keep a cache to track them to avoid duplicate uploads, thereby savings space by no longer unnecessarily uploading files that don't need to be!

Oh, and this also adds a validation check for directories. We won't be supporting those yet.

* patch(plugin/ai): "ai say" command now correctly formats channel message

It previously, uh... sent the payload as is which was not a correct format. ;-;

* feat(plugin/ai): support for function calls w/ assistant API

It's far from perfect but finally usable!

* chore(plugin/ai): document custom models