Feat(env.d.ts)
Creating a env.d.ts file for basic types autocomplete with .env file.

Note: the tsconfig.json file is only created after (at the same time) credentials are asked, that's why the "createEnvTsFile" method is at the end of the "askForDiscordCredentials" one, please, if this is an issue, let me know.