refactor(robo)!: removed legacy process-based runtime
It's finally time. Dev rebuilds using this runtime has been great initially, but our "new" worker thread-based model is much more lightweight and accomplishes the same thing in a superior way.

We had initially left this option to revert in place in case the new model didn't work out. I was surprised to see that such a complex change worked perfectly the first time around!

BREAKING CHANGE:
If you've been using the legacy runtime, it's time to move on.