feat(docs): search bar for templates
Added search bar to the playground to be able to try out different templates.

Please, be aware that I am using Robo so I added it, the reason is that I wanted to dynamically fetch the templates we have so I could display all the most recents ones, please review the package.json carefully especially the commands, If you need more informations about it, just ask me.

PS:
I could not get click events to work on the iFrame and I almost had a breakdown and started crying.