refactor(robo): use djs helpers for option extraction
Works very similarly as before, meaning no breaking changes. Main benefit is being able to let DJS validate options for ya~

(e.g. This is a number, not a string option)

Technically shouldn't be necessary considering we guarantee consistency in extraction and declaration, but it's an added benefit I guess.